+++ tests/test_entity.py
@@ -159,5 +159,5 @@ async def test_generic_wrapped_entity(factory: helper.Factory) -> None:
 def test_custom_required_entities() -> None:
     """Test required parameters from entity definitions."""
     required_parameters = get_required_parameters()
-    assert len(required_parameters) == 72
+    assert len(required_parameters) == 75
     assert check_ignore_parameters_is_clean() is True